[*] refactor le_renderer
move le_renderer cpp facade to le_renderer.hpp so that the templating
does not get invoked everytime that le_renderer is included.

this should speed up compilation and structure the code a bit better.
this change means that every application that wants to use the cpp
facade now must include le_renderer.hpp instead of le_renderer.h, which
might be a bit finnicky...
