+++ modules/le_renderer/private/le_renderer/le_renderer_types.h
@@ -293,293 +293,6 @@ struct le_write_to_image_settings_t {
 };
 
 namespace le {
-using Presentmode = le_swapchain_settings_t::khr_settings_t::Presentmode;
-
-#define BUILDER_IMPLEMENT( builder, method_name, param_type, param, default_value ) \
-	constexpr builder& method_name( param_type param default_value ) {              \
-		self.param = param;                                                         \
-		return *this;                                                               \
-	}
-
-#define P_BUILDER_IMPLEMENT( builder, method_name, param_type, param, default_value ) \
-	constexpr builder& method_name( param_type param default_value ) {                \
-		self->param = param;                                                          \
-		return *this;                                                                 \
-	}
-class RendererInfoBuilder {
-	le_renderer_settings_t   info{};
-	le_renderer_settings_t&  self               = info;
-	le_swapchain_settings_t* swapchain_settings = nullptr;
-	le_window_o*             initial_window     = nullptr;
-
-  public:
-	RendererInfoBuilder( le_window_o* window = nullptr )
-	    : swapchain_settings( info.swapchain_settings )
-	    , initial_window( window ) {
-		if ( window && info.swapchain_settings->type == le_swapchain_settings_t::LE_KHR_SWAPCHAIN ) {
-			info.swapchain_settings->khr_settings.window = window;
-			info.swapchain_settings->type                = le_swapchain_settings_t::Type::LE_KHR_SWAPCHAIN;
-		}
-	}
-
-	class SwapchainInfoBuilder {
-		RendererInfoBuilder&      parent;
-		le_swapchain_settings_t*& self;
-
-	  public:
-		SwapchainInfoBuilder( RendererInfoBuilder& parent_ )
-		    : parent( parent_ )
-		    , self( parent.swapchain_settings ) {
-		}
-
-		P_BUILDER_IMPLEMENT( SwapchainInfoBuilder, setWidthHint, uint32_t, width_hint, = 640 )
-		P_BUILDER_IMPLEMENT( SwapchainInfoBuilder, setHeightHint, uint32_t, height_hint, = 480 )
-		P_BUILDER_IMPLEMENT( SwapchainInfoBuilder, setImagecountHint, uint32_t, imagecount_hint, = 3 )
-		P_BUILDER_IMPLEMENT( SwapchainInfoBuilder, setFormatHint, le::Format, format_hint, = le::Format::eB8G8R8A8Unorm )
-
-		class KhrSwapchainInfoBuilder {
-			SwapchainInfoBuilder& parent;
-
-		  public:
-			KhrSwapchainInfoBuilder( SwapchainInfoBuilder& parent_ )
-			    : parent( parent_ ) {
-				parent.self->init_khr_settings();
-			}
-
-			KhrSwapchainInfoBuilder& setPresentmode( le::Presentmode presentmode_hint = le::Presentmode::eFifo ) {
-				this->parent.parent.swapchain_settings->khr_settings.presentmode_hint = presentmode_hint;
-				return *this;
-			}
-
-			KhrSwapchainInfoBuilder& setWindow( le_window_o* window = nullptr ) {
-				this->parent.parent.swapchain_settings->khr_settings.window = window;
-				return *this;
-			}
-
-			SwapchainInfoBuilder& end() {
-				parent.parent.swapchain_settings->type = le_swapchain_settings_t::Type::LE_KHR_SWAPCHAIN;
-				return parent;
-			}
-		};
-
-		class DirectSwapchainInfoBuilder {
-			SwapchainInfoBuilder& parent;
-
-		  public:
-			DirectSwapchainInfoBuilder( SwapchainInfoBuilder& parent_ )
-			    : parent( parent_ ) {
-				parent.self->init_khr_direct_mode_settings();
-			}
-
-			DirectSwapchainInfoBuilder& setPresentmode( le::Presentmode presentmode_hint = le::Presentmode::eFifo ) {
-				this->parent.parent.swapchain_settings->khr_direct_mode_settings.presentmode_hint = presentmode_hint;
-				return *this;
-			}
-
-			DirectSwapchainInfoBuilder& setDisplayName( char const* display_name = "" ) {
-				this->parent.parent.swapchain_settings->khr_direct_mode_settings.display_name = display_name;
-				return *this;
-			}
-
-			SwapchainInfoBuilder& end() {
-				parent.parent.swapchain_settings->type = le_swapchain_settings_t::Type::LE_DIRECT_SWAPCHAIN;
-				return parent;
-			}
-		};
-
-		class ImgSwapchainInfoBuilder {
-			SwapchainInfoBuilder& parent;
-
-			static constexpr auto default_pipe_cmd = "ffmpeg -r 60 -f rawvideo -pix_fmt rgba -s %dx%d -i - -threads 0 -preset fast -y -pix_fmt yuv420p isl%s.mp4";
-
-		  public:
-			ImgSwapchainInfoBuilder( SwapchainInfoBuilder& parent_ )
-			    : parent( parent_ ) {
-				parent.self->init_img_settings();
-				setPipeCmd();
-			}
-
-			ImgSwapchainInfoBuilder& setPipeCmd( char const* pipe_cmd = default_pipe_cmd ) {
-				parent.parent.swapchain_settings->img_settings.pipe_cmd = pipe_cmd;
-				return *this;
-			}
-
-			SwapchainInfoBuilder& end() {
-				parent.parent.swapchain_settings->type = le_swapchain_settings_t::Type::LE_IMG_SWAPCHAIN;
-				return parent;
-			}
-		};
-
-		DirectSwapchainInfoBuilder mDirectSwapchainInfoBuilder{ *this }; // order matters, last one will be default, because initialisation overwrites.
-		ImgSwapchainInfoBuilder    mImgSwapchainInfoBuilder{ *this };    // order matters, last one will be default, because initialisation overwrites.
-		KhrSwapchainInfoBuilder    mKhrSwapchainInfoBuilder{ *this };    // order matters, last one will be default, because initialisation overwrites.
-
-		KhrSwapchainInfoBuilder& asWindowSwapchain() {
-			this->self->init_khr_settings();
-			return mKhrSwapchainInfoBuilder;
-		}
-
-		ImgSwapchainInfoBuilder& asImgSwapchain() {
-			this->self->init_img_settings();
-			return mImgSwapchainInfoBuilder;
-		}
-
-		DirectSwapchainInfoBuilder& asDirectSwapchain() {
-			this->self->init_khr_direct_mode_settings();
-			return mDirectSwapchainInfoBuilder;
-		}
-
-		RendererInfoBuilder& end() {
-			parent.info.num_swapchain_settings++;
-			parent.swapchain_settings++;
-			return parent;
-		}
-	};
-
-	SwapchainInfoBuilder mSwapchainInfoBuilder{ *this };
-
-	SwapchainInfoBuilder& addSwapchain() {
-		return mSwapchainInfoBuilder;
-	}
-
-	le_renderer_settings_t const& build() {
-		// if an initial window was given and nothing else,
-		// we must make sure that the setting still counts.
-		// if the builder pattern was used, then the window
-		// will have been applied to the first element, and
-		// end() will have been called on the Builder, which
-		// will have incremented `num_swapchain_settings`.
-		if ( initial_window && info.num_swapchain_settings == 0 ) {
-			info.num_swapchain_settings = 1;
-		}
-		return info;
-	}
-};
-
-// ----------------------------------------------------------------------
-
-class ImageSamplerInfoBuilder {
-	le_image_sampler_info_t info{};
-
-	class SamplerInfoBuilder {
-		ImageSamplerInfoBuilder& parent;
-		le_sampler_info_t&       self = parent.info.sampler;
-
-	  public:
-		SamplerInfoBuilder( ImageSamplerInfoBuilder& parent_ )
-		    : parent( parent_ ) {
-		}
-
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setMagFilter, le::Filter, magFilter, = le::Filter::eLinear )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setMinFilter, le::Filter, minFilter, = le::Filter::eLinear )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setMipmapMode, le::SamplerMipmapMode, mipmapMode, = le::SamplerMipmapMode::eLinear )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setAddressModeU, le::SamplerAddressMode, addressModeU, = le::SamplerAddressMode::eClampToBorder )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setAddressModeV, le::SamplerAddressMode, addressModeV, = le::SamplerAddressMode::eClampToBorder )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setAddressModeW, le::SamplerAddressMode, addressModeW, = le::SamplerAddressMode::eRepeat )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setMipLodBias, float, mipLodBias, = 0.f )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setAnisotropyEnable, bool, anisotropyEnable, = false )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setMaxAnisotropy, float, maxAnisotropy, = 0.f )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setCompareEnable, bool, compareEnable, = false )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setCompareOp, le::CompareOp, compareOp, = le::CompareOp::eLess )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setMinLod, float, minLod, = 0.f )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setMaxLod, float, maxLod, = 1.f )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setBorderColor, le::BorderColor, borderColor, = le::BorderColor::eFloatTransparentBlack )
-		BUILDER_IMPLEMENT( SamplerInfoBuilder, setUnnormalizedCoordinates, bool, unnormalizedCoordinates, = false )
-
-		ImageSamplerInfoBuilder& end() {
-			return parent;
-		}
-	};
-
-	class ImageViewInfoBuilder {
-		ImageSamplerInfoBuilder&                       parent;
-		le_image_sampler_info_t::le_image_view_info_t& self = parent.info.imageView;
-
-	  public:
-		ImageViewInfoBuilder( ImageSamplerInfoBuilder& parent_ )
-		    : parent( parent_ ) {
-		}
-
-		BUILDER_IMPLEMENT( ImageViewInfoBuilder, setImage, le_img_resource_handle, imageId, = {} )
-		BUILDER_IMPLEMENT( ImageViewInfoBuilder, setImageViewType, le::ImageViewType, image_view_type, = le::ImageViewType::e2D )
-		BUILDER_IMPLEMENT( ImageViewInfoBuilder, setFormat, le::Format, format, = le::Format::eUndefined )
-		BUILDER_IMPLEMENT( ImageViewInfoBuilder, setBaseArrayLayer, uint32_t, base_array_layer, = 0 )
-		BUILDER_IMPLEMENT( ImageViewInfoBuilder, setLayerCount, uint32_t, layer_count, = 1 )
-
-		ImageSamplerInfoBuilder& end() {
-			return parent;
-		}
-	};
-
-	SamplerInfoBuilder   mSamplerInfoBuilder{ *this };
-	ImageViewInfoBuilder mImageViewInfoBuilder{ *this };
-
-  public:
-	ImageSamplerInfoBuilder()  = default;
-	~ImageSamplerInfoBuilder() = default;
-
-	ImageSamplerInfoBuilder( le_image_sampler_info_t const& info_ )
-	    : info( info_ ) {
-	}
-
-	ImageSamplerInfoBuilder( le_img_resource_handle const& image_resource ) {
-		info.imageView.imageId = image_resource;
-	}
-
-	ImageViewInfoBuilder& withImageViewInfo() {
-		return mImageViewInfoBuilder;
-	}
-
-	SamplerInfoBuilder& withSamplerInfo() {
-		return mSamplerInfoBuilder;
-	}
-
-	le_image_sampler_info_t const& build() {
-		return info;
-	}
-};
-
-class ImageAttachmentInfoBuilder {
-	le_image_attachment_info_t self{};
-
-  public:
-	BUILDER_IMPLEMENT( ImageAttachmentInfoBuilder, setLoadOp, le::AttachmentLoadOp, loadOp, = le::AttachmentLoadOp::eClear )
-	BUILDER_IMPLEMENT( ImageAttachmentInfoBuilder, setStoreOp, le::AttachmentStoreOp, storeOp, = le::AttachmentStoreOp::eStore )
-	BUILDER_IMPLEMENT( ImageAttachmentInfoBuilder, setColorClearValue, le::ClearValue, clearValue, = le_image_attachment_info_t::DefaultClearValueColor )
-	BUILDER_IMPLEMENT( ImageAttachmentInfoBuilder, setDepthStencilClearValue, le::ClearValue, clearValue, = le_image_attachment_info_t::DefaultClearValueDepthStencil )
-
-	le_image_attachment_info_t const& build() {
-		return self;
-	}
-};
-
-class WriteToImageSettingsBuilder {
-	le_write_to_image_settings_t self{};
-
-  public:
-	WriteToImageSettingsBuilder() = default;
-
-	BUILDER_IMPLEMENT( WriteToImageSettingsBuilder, setImageW, uint32_t, image_w, = 0 )
-	BUILDER_IMPLEMENT( WriteToImageSettingsBuilder, setImageH, uint32_t, image_h, = 0 )
-	BUILDER_IMPLEMENT( WriteToImageSettingsBuilder, setImageD, uint32_t, image_d, = 0 )
-	BUILDER_IMPLEMENT( WriteToImageSettingsBuilder, setOffsetX, int32_t, offset_x, = 0 )
-	BUILDER_IMPLEMENT( WriteToImageSettingsBuilder, setOffsetY, int32_t, offset_y, = 0 )
-	BUILDER_IMPLEMENT( WriteToImageSettingsBuilder, setOffsetZ, int32_t, offset_z, = 0 )
-	BUILDER_IMPLEMENT( WriteToImageSettingsBuilder, setArrayLayer, uint32_t, dst_array_layer, = 0 )
-	BUILDER_IMPLEMENT( WriteToImageSettingsBuilder, setDstMiplevel, uint32_t, dst_miplevel, = 0 )
-	BUILDER_IMPLEMENT( WriteToImageSettingsBuilder, setNumMiplevels, uint32_t, num_miplevels, = 1 )
-
-	constexpr le_write_to_image_settings_t const& build() const {
-		return self;
-	}
-};
-
-#undef BUILDER_IMPLEMENT
-#undef P_BUILDER_IMPLEMENT
-
-} // namespace le
-
-namespace le {
 enum class RayTracingShaderGroupType : uint32_t {
 	eRayGen = 0,
 	eTrianglesHitGroup,
